pyproject.toml
[project]
name = "agent-sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "cloudscraper>=1.2.71",
    "fake-useragent>=2.2.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.3.7",
    "langchain-postgres>=0.0.15",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-cli[inmem]>=0.4.2",
    "pandas>=2.3.2",
    "playwright>=1.55.0",
    "psycopg[binary,pool]>=3.2.10",
    "pytest-playwright>=0.7.1",
    "pyyaml>=6.0.2",
    "selenium>=4.35.0",
    "streamlit>=1.49.1",
    "tabulate>=0.9.0",
    "undetected-chromedriver>=3.5.5",
]
